Change the name of the Rails Application
I am trying to emphasies and reenforce the seperation between 'Application' and
'Delivery Mechanism'. It is my intention that the application (business/domain
logic) will live in a seperate gem which will be included in the Rails app's
Gemfile and required by those pieces of rails that depend on it.

Please refer to Robert Martin's "Architecture the Lost Years" talk for details.