Move all of Lendger under lib/
I want to continue focusing on seperation of concerns. As such, and given that
I want the Framework (Rails) to depend on the Application (Lendger), I have
moved anything that is exclusively a 'Business' concern under the lib/lendger/
directory.

I also created a spec/lendger subdirectory and moved Lendger specs under that.